requirements/pyproject.toml
[tool.poetry]
name = "sibdev-test-2"
version = "0.1.0"
description = "Test task for sibdev company."
authors = ["UncleVasya <unclevasyaai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.4"
gunicorn = "^21.2.0"
djangorestframework = "^3.14.0"
django-redis = "^5.3.0"
whitenoise = "^6.5.0"
faker = "^19.3.1"
coverage = "^7.3.0"
fakeredis = "^2.18.0"
factory-boy = "^3.3.0"
psycopg2-binary = "^2.9.7"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.3.0"
tqdm = "^4.66.1"
redis = "^5.0.0"
celery = "^5.3.4"
django-environ = "^0.11.2"
django-filter = "^23.2"
drf-spectacular = "^0.26.4"
django-post-office = "^3.7.1"
django-solo = "^2.1.0"
responses = "^0.23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
